+++ src/libtrx/include/libtrx/game/clock/timer.h
@@ -1,29 +1,17 @@
 #pragma once
 
-#include <stdbool.h>
-#include <stddef.h>
-#include <stdint.h>
+typedef enum {
+    CLOCK_TIMER_REAL,
+    CLOCK_TIMER_SIM,
+} CLOCK_TIMER_TYPE;
 
 typedef struct {
-    double prev_counter;
-    int32_t prev_fps;
+    double ref;
+    CLOCK_TIMER_TYPE type;
 } CLOCK_TIMER;
 
-void Clock_ResetTimer(CLOCK_TIMER *timer);
-
-// Check how many logical frames passed since the last run of this function.
-// The `timer` argument is used to store the previous frame count and should
-// not be reused between contexts this function is called with.
-double Clock_GetElapsedLogicalFrames(CLOCK_TIMER *timer);
-double Clock_GetElapsedDrawFrames(CLOCK_TIMER *timer);
-double Clock_GetElapsedMilliseconds(CLOCK_TIMER *timer);
-
-// Similar logic to Clock_GetElapsedLogicalFrames, except reset the timer only
-// if there are enough elapsed frames.
-bool Clock_CheckElapsedLogicalFrames(CLOCK_TIMER *timer, int32_t wait);
-bool Clock_CheckElapsedDrawFrames(CLOCK_TIMER *timer, int32_t wait);
-bool Clock_CheckElapsedMilliseconds(CLOCK_TIMER *timer, int32_t wait);
-
-// The same as Clock_CheckElapsedMilliseconds, except does not scale the result
-// by the turbo cheat multiplier.
-bool Clock_CheckElapsedRawMilliseconds(CLOCK_TIMER *timer, int32_t how_often);
+void ClockTimer_Sync(CLOCK_TIMER *timer);
+double ClockTimer_PeekElapsed(const CLOCK_TIMER *timer);
+double ClockTimer_TakeElapsed(CLOCK_TIMER *timer);
+bool ClockTimer_CheckElapsed(const CLOCK_TIMER *timer, double sec);
+bool ClockTimer_CheckElapsedAndTake(CLOCK_TIMER *timer, double sec);