+++ src/tr2/game/input.c
@@ -11,11 +11,11 @@
 #include <libtrx/game/input/backends/controller.h>
 #include <libtrx/game/input/backends/keyboard.h>
 
-#define DELAY_FRAMES 12
-#define HOLD_FRAMES 3
+#define DELAY_TIME 0.4
+#define HOLD_TIME 0.1
 
-static int32_t m_HoldBack = 0;
-static int32_t m_HoldForward = 0;
+static CLOCK_TIMER m_HoldBackTimer = { .type = CLOCK_TIMER_REAL };
+static CLOCK_TIMER m_HoldForwardTimer = { .type = CLOCK_TIMER_REAL };
 
 static INPUT_STATE M_GetDebounced(INPUT_STATE input);
 static void M_UpdateFromBackend(
@@ -26,26 +26,33 @@ static INPUT_STATE M_GetDebounced(const INPUT_STATE input)
     INPUT_STATE result;
     result.any = input.any & ~g_OldInputDB.any;
 
-    const int32_t frame = Clock_GetLogicalFrame();
+    // Allow holding down key to move faster
     if (input.forward || !input.back) {
-        m_HoldBack = 0;
-    } else if (input.back && m_HoldBack == 0) {
-        m_HoldBack = frame;
-    } else if (input.back && frame - m_HoldBack >= DELAY_FRAMES + HOLD_FRAMES) {
+        m_HoldBackTimer.ref = 0.0;
+    } else if (input.back && m_HoldBackTimer.ref == 0.0) {
+        ClockTimer_Sync(&m_HoldBackTimer);
+    } else if (
+        input.back
+        && ClockTimer_CheckElapsedAndTake(
+            &m_HoldBackTimer, DELAY_TIME + HOLD_TIME)) {
         result.back = 1;
         result.menu_down = 1;
-        m_HoldBack = frame - DELAY_FRAMES;
+        ClockTimer_Sync(&m_HoldBackTimer);
+        m_HoldBackTimer.ref -= DELAY_TIME;
     }
 
     if (!input.forward || input.back) {
-        m_HoldForward = 0;
-    } else if (input.forward && m_HoldForward == 0) {
-        m_HoldForward = frame;
+        m_HoldForwardTimer.ref = 0.0;
+    } else if (input.forward && m_HoldForwardTimer.ref == 0.0) {
+        ClockTimer_Sync(&m_HoldForwardTimer);
     } else if (
-        input.forward && frame - m_HoldForward >= DELAY_FRAMES + HOLD_FRAMES) {
+        input.forward
+        && ClockTimer_CheckElapsed(
+            &m_HoldForwardTimer, DELAY_TIME + HOLD_TIME)) {
         result.forward = 1;
         result.menu_up = 1;
-        m_HoldForward = frame - DELAY_FRAMES;
+        ClockTimer_Sync(&m_HoldForwardTimer);
+        m_HoldForwardTimer.ref -= DELAY_TIME;
     }
 
     g_OldInputDB = input;